transcode file contents on reading file
file content is iterated twice (on determining chunk ranges and on
printing the results). To avoid transcoding twice, this patch transcodes
the file content once on reading it from file. However this has a down
side that transcoding the entire file is redundant. In the future,
on-demand transcoding may be considered.